refactor: TMS-2208 clean up create tenant code and ui
Rename the CreateTenantDialog
